backend/pyproject.toml
[project]
name = "atlas-assistant"
version = "0.0.1"
description = "Africa Agriculture Adaptation Atlas Assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "langchain-chroma>=1.0.0",
    "langchain-mistralai>=1.0.0",
    "langchain>=1.0.2",
    "langgraph>=1.0.1",
    "mistralai>=1.9.11",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "basedpyright>=1.32.0",
    "chainlit>=2.8.3",
    "fastapi[standard]>=0.116.2",
    "httpx>=0.28.1",
    "prek>=0.2.10",
    "pytest>=8.4.2",
]

[tool.basedpyright]
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportMissingTypeStubs = false
reportAny = false
reportExplicitAny = false

[tool.pytest.ini_options]
markers = "integration: Run integration tests, which require an LLM API key"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
